# folkify_BE/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FOLKIFY Backend API",
    "version": "1.0.0",
    "description": "REST API for FOLKIFY - Vietnamese Folk Music Learning Platform. Supports authentication, premium subscriptions (FREE/BASIC/PRO), lessons with access control, AI grading service, and admin dashboard.",
    "contact": {
      "name": "FOLKIFY Team",
      "email": "support@folkify.com"
    },
    "license": {
      "name": "ISC"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.folkify.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Access Token (15 minutes expiration)"
      }
    },
    "schemas": {
...
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 20
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 5
          }
        }
      },
...
      "Lesson": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "instrument_id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "description": "Duration in minutes"
          },
          "level": {
            "type": "string",
            "enum": [
              "Beginner",
              "Intermediate",
              "Advanced"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ]
          },
          "is_premium": {
            "type": "boolean"
          },
          "youtube_embed_url": {
            "type": "string",
            "nullable": true
          },
          "video_thumb": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "xp": {
            "type": "integer"
          },
          "order_index": {
            "type": "integer"
          },
          "steps": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "nullable": true
          },
          "tips": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "has_access": {
            "type": "boolean",
            "description": "Whether the current user has access to this lesson"
          },
          "requires_premium": {
            "type": "boolean",
            "description": "Whether this lesson requires premium subscription"
          },
          "completed": {
            "type": "boolean",
            "description": "Whether the current user has completed this lesson"
          },
          "progress_percentage": {
            "type": "integer",
            "description": "User progress percentage (0-100)"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/lessons/search": {
      "get": {
        "summary": "Search lessons",
        "tags": [
          "Lessons"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string"
            },
            "description": "Search term"
          },
          {
            "in": "query",
            "name": "level",
            "schema": {
              "type": "string",
              "enum": [
                "Beginner",
                "Intermediate",
                "Advanced"
              ]
            }
          },
          {
            "in": "query",
            "name": "instrument_id",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Lesson"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/lessons/{id}": {
      "get": {
        "summary": "Get lesson by ID with access control",
        "tags": [
          "Lessons"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lesson details with access information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Lesson"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied - Premium required"
          },
          "404": {
            "description": "Lesson not found"
          }
        }
      }
    },
    "/api/lessons/{id}/complete": {
      "post": {
        "summary": "Mark lesson as completed",
        "tags": [
          "Lessons"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lesson completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "xp_earned": {
                          "type": "integer"
                        },
                        "new_total_xp": {
                          "type": "integer"
                        },
                        "new_level": {
                          "type": "integer"
                        },
                        "level_up": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Access denied"
          },
          "404": {
            "description": "Lesson not found"
          }
        }
      }
    },
...
  }
}